Refactor FaceEngine and MiniFasNetEngine for Liveness Detection and Embedding Extraction
- Updated FaceEngine interface to support loading multiple models and added LivenessResult data class.
- Modified MiniFasNetEngine to handle separate sessions for liveness detection and embedding extraction.
- Changed method signatures and implementations to accommodate new model loading and processing logic.
- Updated SyncRepositoryImpl and SyncService to reflect changes in data entities and methods.
- Removed outdated AttendanceLogicTest and SyncServiceTest classes.
- Updated Gradle and plugin versions for improved compatibility and performance.
- Added ONNX model files for ArcFace and MiniFasNet to assets.

+++ app/src/main/java/com/smkn2malinau/absensi/data/remote/DTOs.kt
@@ -2,14 +2,13 @@ package com.smkn2malinau.absensi.data.remote
 
 import com.google.gson.annotations.SerializedName
 
-// --- DTOs untuk Absensi Sync ---
 data class SyncAbsensiRequest(
     @SerializedName("records") val records: List<AbsensiRecordDto>
 )
 
 data class AbsensiRecordDto(
     @SerializedName("record_id") val recordId: String,
-    @SerializedName("siswa_id") val siswaId: Long,
+    @SerializedName("siswa_id") val siswaId: Int,
     @SerializedName("tanggal") val tanggal: String,
     @SerializedName("type") val type: String,
     @SerializedName("jam_aktual") val jamAktual: String,
@@ -20,27 +19,25 @@ data class AbsensiRecordDto(
 
 data class SyncAbsensiResponse(
     @SerializedName("status") val status: String,
-    @SerializedName("diterima") val diterima: List<String>, // list of record_id yang berhasil
+    @SerializedName("diterima") val diterima: List<String>,
     @SerializedName("duplikat") val duplikat: List<String>,
     @SerializedName("gagal") val gagal: List<String>
 )
 
-// --- DTOs untuk Embedding Sync ---
 data class EmbeddingSyncResponse(
     @SerializedName("siswa") val siswaList: List<SiswaEmbeddingDto>
 )
 
 data class SiswaEmbeddingDto(
-    @SerializedName("siswa_id") val siswaId: Long,
+    @SerializedName("siswa_id") val siswaId: Int,
     @SerializedName("nis") val nis: String,
     @SerializedName("nama") val nama: String,
     @SerializedName("kelas") val kelas: String,
     @SerializedName("embedding_base64") val embeddingBase64: String?,
     @SerializedName("model_version") val modelVersion: String,
-    @SerializedName("aktif") val aktif: Boolean = true // PRD 9.2: field aktif
+    @SerializedName("aktif") val aktif: Boolean = true
 )
 
-// --- DTOs untuk Jadwal Efektif ---
 data class JadwalEfektifResponse(
     @SerializedName("jadwal") val jadwalList: List<JadwalDto>
 )
@@ -54,22 +51,20 @@ data class JadwalDto(
     @SerializedName("sumber") val sumber: String
 )
 
-// --- DTOs untuk Dispensasi Aktif ---
 data class DispensasiAktifResponse(
     @SerializedName("dispensasi") val dispensasiList: List<DispensasiDto>
 )
 
 data class DispensasiDto(
-    @SerializedName("siswa_id") val siswaId: Long,
+    @SerializedName("siswa_id") val siswaId: Int,
     @SerializedName("tanggal") val tanggal: String,
     @SerializedName("jenis") val jenis: String,
     @SerializedName("kategori") val kategori: String,
     @SerializedName("alasan") val alasan: String?
 )
 
-// --- DTOs untuk Push Override Jadwal (PRD 9.1) ---
 data class PushOverrideRequest(
-    @SerializedName("client_id") val clientId: String, // UUID
+    @SerializedName("client_id") val clientId: String,
     @SerializedName("tanggal") val tanggal: String,
     @SerializedName("kelas") val kelas: String?,
     @SerializedName("jam_masuk") val jamMasuk: String,
@@ -78,11 +73,10 @@ data class PushOverrideRequest(
 )
 
 data class PushOverrideResponse(
-    @SerializedName("status") val status: String, // "ok" | "ditolak"
+    @SerializedName("status") val status: String,
     @SerializedName("pesan") val pesan: String?
 )
 
-// --- DTOs untuk Lapor Kesehatan Device (PRD 9.3) ---
 data class HealthReportRequest(
     @SerializedName("jadwal_jam_lalu") val jadwalJamLalu: Long,
     @SerializedName("dispensasi_jam_lalu") val dispensasiJamLalu: Long
